chore: 로그인 여부 및 멤버쉽 api 함수 분리
